+++ src/components/snippetsModal/snippetItem.tsx
@@ -1,5 +1,7 @@
 import * as React from 'react';
-import { Text, View, Button } from 'react-native';
+import { useIntl } from 'react-intl';
+import { Alert, Text, View } from 'react-native';
+import { useSnippetDeleteMutation } from '../../providers/queries';
 import IconButton from '../iconButton';
 import styles from './snippetsModalStyles';
 
@@ -9,10 +11,32 @@ interface SnippetItemProps {
   body: string;
   index: number;
   onEditPress: () => void;
-  onRemovePress: () => void;
 }
 
-const SnippetItem = ({ id, title, body, index, onEditPress, onRemovePress }: SnippetItemProps) => {
+const SnippetItem = ({ id, title, body, index, onEditPress }: SnippetItemProps) => {
+  const intl = useIntl();
+  const snippetsDeleteMutation = useSnippetDeleteMutation();
+
+  const _onRemovePress = () => {
+    //asks for remvoe confirmation and run remove routing upon confirming
+    if (id) {
+      Alert.alert(
+        intl.formatMessage({ id: 'snippets.title_remove_confirmation' }),
+        intl.formatMessage({ id: 'snippets.message_remove_confirmation' }),
+        [
+          {
+            text: intl.formatMessage({ id: 'snippets.btn_cancel' }),
+            style: 'cancel',
+          },
+          {
+            text: intl.formatMessage({ id: 'snippets.btn_confirm' }),
+            onPress: () => snippetsDeleteMutation.mutate(id),
+          },
+        ],
+      );
+    }
+  };
+
   return (
     <View style={[styles.itemWrapper, index % 2 !== 0 && styles.itemWrapperGray]}>
       <View style={styles.itemHeader}>
@@ -30,9 +54,10 @@ const SnippetItem = ({ id, title, body, index, onEditPress, onRemovePress }: Sni
             <IconButton
               iconStyle={styles.itemIcon}
               style={styles.itemIconWrapper}
+              isLoading={snippetsDeleteMutation.isLoading}
               iconType="MaterialCommunityIcons"
               name="delete"
-              onPress={onRemovePress}
+              onPress={_onRemovePress}
               size={20}
             />
           </>
